Create LinkedIssue for endpoint "/projects/:id/issues/:issue_iid/links"
