+++ src/main/java/org/gitlab4j/api/IssuesApi.java
@@ -17,6 +17,7 @@ import org.gitlab4j.api.models.IssueLink;
 import org.gitlab4j.api.models.IssuesStatistics;
 import org.gitlab4j.api.models.IssuesStatisticsFilter;
 import org.gitlab4j.api.models.LinkType;
+import org.gitlab4j.api.models.LinkedIssue;
 import org.gitlab4j.api.models.MergeRequest;
 import org.gitlab4j.api.models.Participant;
 import org.gitlab4j.api.models.TimeStats;
@@ -834,7 +835,7 @@ public class IssuesApi extends AbstractApi implements Constants {
      * @return a list of related issues of a given issue, sorted by the relationship creation datetime (ascending)
      * @throws GitLabApiException if any exception occurs
      */
-    public List<Issue> getIssueLinks(Object projectIdOrPath, Long issueIid) throws GitLabApiException {
+    public List<LinkedIssue> getIssueLinks(Object projectIdOrPath, Long issueIid) throws GitLabApiException {
         return (getIssueLinks(projectIdOrPath, issueIid, getDefaultPerPage()).all());
     }
 
@@ -852,8 +853,8 @@ public class IssuesApi extends AbstractApi implements Constants {
      * @return a Pager of related issues of a given issue, sorted by the relationship creation datetime (ascending)
      * @throws GitLabApiException if any exception occurs
      */
-    public Pager<Issue> getIssueLinks(Object projectIdOrPath, Long issueIid, int itemsPerPage) throws GitLabApiException {
-        return (new Pager<Issue>(this, Issue.class, itemsPerPage, null,
+    public Pager<LinkedIssue> getIssueLinks(Object projectIdOrPath, Long issueIid, int itemsPerPage) throws GitLabApiException {
+        return (new Pager<LinkedIssue>(this, LinkedIssue.class, itemsPerPage, null,
                 "projects", getProjectIdOrPath(projectIdOrPath), "issues", issueIid, "links"));
     }
 
@@ -870,7 +871,7 @@ public class IssuesApi extends AbstractApi implements Constants {
      * @return a Stream of related issues of a given issue, sorted by the relationship creation datetime (ascending)
      * @throws GitLabApiException if any exception occurs
      */
-    public Stream<Issue> getIssueLinksStream(Object projectIdOrPath, Long issueIid) throws GitLabApiException {
+    public Stream<LinkedIssue> getIssueLinksStream(Object projectIdOrPath, Long issueIid) throws GitLabApiException {
         return (getIssueLinks(projectIdOrPath, issueIid, getDefaultPerPage()).stream());
     }
