+++ src/main/java/org/gitlab4j/api/models/Issue.java
@@ -1,18 +1,11 @@
 
 package org.gitlab4j.api.models;
 
-import java.util.Date;
-
 import org.gitlab4j.api.utils.JacksonJson;
 
 public class Issue extends AbstractIssue {
 
     private Boolean subscribed;
-
-    private Long issueLinkId;
-    private LinkType linkType;
-    private Date linkCreatedAt;
-    private Date linkUpdatedAt;
 
     public Boolean getSubscribed() {
         return subscribed;
@@ -20,38 +13,6 @@ public class Issue extends AbstractIssue {
 
     public void setSubscribed(Boolean subscribed) {
         this.subscribed = subscribed;
-    }
-
-    public Long getIssueLinkId() {
-        return issueLinkId;
-    }
-
-    public void setIssueLinkId(Long issueLinkId) {
-        this.issueLinkId = issueLinkId;
-    }
-
-    public LinkType getLinkType() {
-        return linkType;
-    }
-
-    public void setLinkType(LinkType linkType) {
-        this.linkType = linkType;
-    }
-
-    public Date getLinkCreatedAt() {
-        return linkCreatedAt;
-    }
-
-    public void setLinkCreatedAt(Date linkCreatedAt) {
-        this.linkCreatedAt = linkCreatedAt;
-    }
-
-    public Date getLinkUpdatedAt() {
-        return linkUpdatedAt;
-    }
-
-    public void setLinkUpdatedAt(Date linkUpdatedAt) {
-        this.linkUpdatedAt = linkUpdatedAt;
     }
 
     @Override
