+++ src/main/java/org/gitlab4j/api/models/LinkedIssue.java
@@ -1,0 +1,51 @@
+
+package org.gitlab4j.api.models;
+
+import java.util.Date;
+
+import org.gitlab4j.api.utils.JacksonJson;
+
+public class LinkedIssue extends AbstractIssue {
+
+    private Long issueLinkId;
+    private LinkType linkType;
+    private Date linkCreatedAt;
+    private Date linkUpdatedAt;
+
+    public Long getIssueLinkId() {
+        return issueLinkId;
+    }
+
+    public void setIssueLinkId(Long issueLinkId) {
+        this.issueLinkId = issueLinkId;
+    }
+
+    public LinkType getLinkType() {
+        return linkType;
+    }
+
+    public void setLinkType(LinkType linkType) {
+        this.linkType = linkType;
+    }
+
+    public Date getLinkCreatedAt() {
+        return linkCreatedAt;
+    }
+
+    public void setLinkCreatedAt(Date linkCreatedAt) {
+        this.linkCreatedAt = linkCreatedAt;
+    }
+
+    public Date getLinkUpdatedAt() {
+        return linkUpdatedAt;
+    }
+
+    public void setLinkUpdatedAt(Date linkUpdatedAt) {
+        this.linkUpdatedAt = linkUpdatedAt;
+    }
+
+    @Override
+    public String toString() {
+        return (JacksonJson.toJsonString(this));
+    }
+}
