+++ src/test/resources/org/gitlab4j/api/epic-issue.json
@@ -59,7 +59,6 @@
     "award_emoji": "http://localhost:3001/api/v4/projects/8/issues/6/award_emoji",
     "project": "http://localhost:3001/api/v4/projects/8"
   },
-  "subscribed": true,
   "epic_issue_id": 2,
   "relative_position": 55
 }
