+++ src/test/resources/org/gitlab4j/api/linked-issues.json
@@ -19,7 +19,6 @@
       "username" : "eileen.lowe"
     },
     "updated_at" : "2016-01-07T12:44:33.959Z",
-    "subscribed" : true,
     "user_notes_count": 0,
     "web_url": "http://example.com/example/example/issues/14",
     "confidential": false,
